add test helper module with a function to create wallets and chains
