+++ test/data.js
@@ -1,21 +1,7 @@
+const helpers = require("./helpers");
 const assert = require("assert");
-const crypto = require("crypto");
 const Tagged = require("cbor/lib/tagged");
 const Tupelo = require("../lib/tupelo");
-const TUPELO_HOST = process.env.TUPELO_RPC_HOST || 'localhost:50051';
-
-const createWalletWithChain = async () => {
-  const wallet = Tupelo.connect(TUPELO_HOST, {
-    walletName: crypto.randomBytes(32).toString('hex'),
-    passPhrase: "test",
-  });
-  await wallet.register();
-  let resp = await wallet.generateKey();
-  const walletKey = resp.keyAddr;
-  assert.equal(42, walletKey.length);
-  let {chainId,} = await wallet.createChainTree(walletKey);
-  return {wallet: wallet, walletKey: walletKey, chainId: chainId};
-};
 
 const itRequires = (version) => {
   const curVer = process.env.TUPELO_VERSION || "";
@@ -62,7 +48,7 @@ describe("setting and retrieving data", function() {
   };
 
   it("can set and retrieve keys with basic values", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     for ([key, val] of Object.entries(basicTypes)) {
       resp = await wallet.setData(chainId, walletKey, key, val);
@@ -92,7 +78,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve keys with array values", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     for ([key, val] of Object.entries(basicTypes)) {
       resp = await wallet.setData(chainId, walletKey, "path/to/" + key, [val, val, val]);
@@ -110,7 +96,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve keys with object values", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     for ([key, val] of Object.entries(basicTypes)) {
       resp = await wallet.setData(chainId, walletKey, "path/to/" + key, {key: val});
@@ -128,7 +114,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve the root data object", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "/", basicTypes);
     assert.notEqual(resp.tip, null);
@@ -139,7 +125,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve sibling keys", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "parent/sibling1", "val1");
     assert.notEqual(resp.tip, null);
@@ -152,7 +138,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve first cousin keys", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "parent/sibling1/child", "val1");
     assert.notEqual(resp.tip, null);
@@ -167,7 +153,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve second cousin keys", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "parent/sibling1/child/anotherChild", "val1");
     assert.notEqual(resp.tip, null);
@@ -182,7 +168,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve basic value on ancestor with existing descendant", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "parent/sibling1/child", "val1");
     assert.notEqual(resp.tip, null);
@@ -199,7 +185,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can set and retrieve descendant with with existing ancestor", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "parent", {name: "val1"});
     assert.notEqual(resp.tip, null);
@@ -219,7 +205,7 @@ describe("setting and retrieving data", function() {
   });
 
   it("can overwrite keys", async ()=> {
-    let {wallet, walletKey, chainId} = await createWalletWithChain();
+    let {wallet, walletKey, chainId} = await helpers.createWalletWithChain();
 
     resp = await wallet.setData(chainId, walletKey, "/", {stableKey: "val1", changingKey: "val2"});
     assert.notEqual(resp.tip, null);
